LPS-77521 Refactor SoyJavaScriptRenderer to become SoyJavaScriptRendererUtil and move to its usage.
