+++ modules/apps/foundation/frontend-taglib/frontend-taglib-soy/src/main/resources/com/liferay/frontend/taglib/soy/internal/util/dependencies/bootstrap.js.tpl
